Collapse sequences of tokens by condition

Concatenates sequences of tokens in the tidy text dataset, while grouping them by an expression.

Usage

collapse_tokens(tbl, condition, .collapse = "")

Arguments

tbl

A tidy text dataset.

condition

A logical expression.

.collapse

String with which tokens are concatenated.

Value

data.frame.

Details

Note that this function drops all columns except but 'token' and columns for grouping sequences. So, the returned data.frame has only 'doc_id', 'sentence_id', 'token_id', and 'token' columns.

Examples

if (FALSE) {
df <- tokenize(
  data.frame(
    doc_id = seq_along(audubon::polano[5:8]),
    text = audubon::polano[5:8]
  )
) |>
  prettify(col_select = "POS1")

collapse_tokens(df, POS1 == "\u540d\u8a5e")
}