Tokenize sentence for data.frame

Tokenize sentence for data.frame

Usage

tokenize(
  tbl,
  text_field = "text",
  docid_field = "doc_id",
  sys_dic = "",
  user_dic = "",
  split = FALSE
)

Arguments

tbl

A data.frame.

text_field

String or symbol; column name where to get texts.

docid_field

String or symbol; column name where to get identifiers of texts.

sys_dic

Character scalar; path to the system dictionary for mecab. Note that the system dictionary is expected to be compiled with UTF-8, not Shift-JIS or other encodings.

user_dic

Character scalar; path to the user dictionary for mecab.

split

Logical. If true, the function internally split the sentence into sub-sentences using stringi::stri_split_boudaries(type = "sentence").

Value

data.frame.

Examples

df <- tokenize(
  data.frame(
    doc_id = seq_len(length(audubon::polano[5:8])),
    text = audubon::polano[5:8]
  )
)
head(df)
#>   doc_id sentence_id token_id    token
#> 1      1           1        1     その
#> 2      1           1        2     ころ
#> 3      1           1        3 わたくし
#> 4      1           1        4       は
#> 5      1           1        5       、
#> 6      1           1        6 モリーオ
#>                                             feature
#> 1                   連体詞,*,*,*,*,*,その,ソノ,ソノ
#> 2         名詞,非自立,副詞可能,*,*,*,ころ,コロ,コロ
#> 3 名詞,代名詞,一般,*,*,*,わたくし,ワタクシ,ワタクシ
#> 4                      助詞,係助詞,*,*,*,*,は,ハ,ワ
#> 5                        記号,読点,*,*,*,*,、,、,、
#> 6                     名詞,固有名詞,地域,一般,*,*,*